Implement Bug Fixes and Cleaned Up Code
- reworked activity_main.xml to only hold recyclerview
- moved necessary ui/ux to fragment_home.xml
- added function and xml to show "no items to show" message when no items can be shown in home view
- unified project to use resources from dimens file for standardized sizing
