+++ app/src/main/res/drawable/edit_text_background.xml
@@ -2,15 +2,15 @@
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:state_enabled="true" android:state_focused="true">
         <shape android:shape="rectangle">
-            <stroke android:width="1dp" android:color="@color/design_default_color_primary" />
-            <corners android:radius="8dp"/>
+            <stroke android:width="@dimen/xxxs" android:color="@color/design_default_color_primary" />
+            <corners android:radius="@dimen/xs"/>
         </shape>
     </item>
 
     <item android:state_enabled="true">
         <shape android:shape="rectangle">
-            <stroke android:width="1dp" android:color="@color/black"/>
-            <corners android:radius="6dp"/>
+            <stroke android:width="@dimen/xxxs" android:color="@color/black"/>
+            <corners android:radius="@dimen/xxs"/>
         </shape>
     </item>
 </selector>
