+++ app/src/main/res/drawable/ic_add.xml
@@ -1,4 +1,9 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#000000" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+<vector xmlns:android="http://schemas.android.com/apk/res/android" 
+    android:height="@dimen/xxxl" 
+    android:tint="#000000" 
+    android:viewportHeight="24" 
+    android:viewportWidth="24" 
+    android:width="@dimen/xxxl">
       
     <path android:fillColor="@android:color/white" android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
